feat: apply review suggestions
